RBAC-main/backend/requirements.txt
# ============================================================================
# Core FastAPI & Server
# ============================================================================
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# ============================================================================
# Validation & Settings
# ============================================================================
pydantic==2.6.4
email-validator==2.1.1
pydantic-settings==2.2.1

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ============================================================================
# Testing
# ============================================================================
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0

# ============================================================================
# Data Processing & Export
# ============================================================================
pandas==2.2.1
openpyxl==3.1.2
python-magic==0.4.27
PyPDF2==3.0.1

# ============================================================================
# TopicLens - Web Scraping & Content Analysis
# ============================================================================
requests==2.32.3
beautifulsoup4==4.12.3
fake-useragent==1.5.1
lxml==5.2.2
praw==7.7.1
DrissionPage==4.1.1.2
SpeechRecognition==3.15.2
pydub==0.25.1

# ============================================================================
# TopicLens - Background Jobs
# ============================================================================
celery==5.4.0
redis==5.0.6

# ============================================================================
# TopicLens - AI/LLM (Hugging Face)
# ============================================================================
transformers>=4.36.0
# Note: torch should be installed separately with CUDA support in Dockerfile
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=4.25.0
